[editor] packages renamed [impact] dependencies added
